test: move route macro tests from lib.rs into route/mod.rs
