+++ rapina-macros/src/route/mod.rs
@@ -58,7 +58,7 @@ impl syn::parse::Parse for RouteAttr {
 }
 
 /// Join a group prefix with a route path at compile time.
-pub(crate) fn join_paths(prefix: &str, path: &str) -> String {
+fn join_paths(prefix: &str, path: &str) -> String {
     let prefix = prefix.trim_end_matches('/');
     if path.is_empty() || path == "/" {
         if prefix.is_empty() {
@@ -78,7 +78,7 @@ pub(crate) fn route_macro(method: &str, attr: TokenStream, item: TokenStream) ->
     route_macro_core(method, attr.into(), item.into()).into()
 }
 
-pub(crate) fn route_macro_core(
+fn route_macro_core(
     method: &str,
     attr: proc_macro2::TokenStream,
     item: proc_macro2::TokenStream,
@@ -455,5 +455,723 @@ pub(crate) fn route_macro_core(
                 register: #register_fn_name,
             }
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use quote::quote;
+
+    #[test]
+    fn test_generates_struct_with_handler_impl() {
+        let path = quote!("/");
+        let input = quote! {
+            async fn hello() -> &'static str {
+                "Hello, Rapina!"
+            }
+        };
+
+        let output = route_macro_core("GET", path, input);
+        let output_str = output.to_string();
+
+        // Check struct is generated
+        assert!(output_str.contains("struct hello"));
+        // Check Handler impl is generated
+        assert!(output_str.contains("impl rapina :: handler :: Handler for hello"));
+        // Check NAME constant
+        assert!(output_str.contains("const NAME"));
+        assert!(output_str.contains("\"hello\""));
+    }
+
+    #[test]
+    fn test_generates_handler_with_extractors() {
+        let path = quote!("/users/:id");
+        let input = quote! {
+            async fn get_user(id: rapina::extract::Path<u64>) -> String {
+                format!("{}", id.into_inner())
+            }
+        };
+
+        let output = route_macro_core("GET", path, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("struct get_user"));
+        // Single arg is last arg — uses FromRequest (blanket impl handles parts-only)
+        assert!(output_str.contains("FromRequest"));
+        // Single arg should NOT destructure request into parts
+        assert!(!output_str.contains("into_parts"));
+    }
+
+    #[test]
+    fn test_function_with_multiple_extractors() {
+        let path = quote!("/users");
+        let input = quote! {
+            async fn create_user(
+                id: rapina::extract::Path<u64>,
+                body: rapina::extract::Json<String>
+            ) -> String {
+                "created".to_string()
+            }
+        };
+
+        let output = route_macro_core("POST", path, input);
+        let output_str = output.to_string();
+
+        // Check struct is generated
+        assert!(output_str.contains("struct create_user"));
+        // Check both extractors are handled
+        assert!(output_str.contains("FromRequestParts"));
+        assert!(output_str.contains("FromRequest"));
+    }
+
+    #[test]
+    fn test_two_body_extractors_no_macro_panic() {
+        // With positional convention, the macro does NOT panic for multiple body consumers.
+        // Instead, it generates code where the first Json is bounded by FromRequestParts
+        // (which it doesn't implement), so the compiler catches it at type-check time.
+        let path = quote!("/users");
+        let input = quote! {
+            async fn handler(
+                body1: rapina::extract::Json<String>,
+                body2: rapina::extract::Json<String>
+            ) -> String {
+                "ok".to_string()
+            }
+        };
+
+        // Should NOT panic — macro expansion succeeds, compiler catches the error later
+        let output = route_macro_core("POST", path, input);
+        let output_str = output.to_string();
+
+        // First arg gets FromRequestParts (will fail at compile time since Json doesn't impl it)
+        assert!(output_str.contains("FromRequestParts"));
+        // Last arg gets FromRequest
+        assert!(output_str.contains("FromRequest"));
+    }
+
+    #[test]
+    fn test_custom_type_name_not_misclassified() {
+        // UserPathInfo contains "Path" but should NOT be routed to FromRequestParts
+        // Positional convention: single (last) arg always uses FromRequest
+        let path = quote!("/users");
+        let input = quote! {
+            async fn handler(info: UserPathInfo) -> String {
+                "ok".to_string()
+            }
+        };
+
+        let output = route_macro_core("POST", path, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("FromRequest"));
+        assert!(!output_str.contains("FromRequestParts"));
+    }
+
+    #[test]
+    fn test_multiple_parts_only_extractors_positional() {
+        // All parts-only extractors: first N-1 use FromRequestParts, last uses FromRequest
+        let path = quote!("/users/:id");
+        let input = quote! {
+            async fn handler(
+                id: rapina::extract::Path<u64>,
+                query: rapina::extract::Query<Params>,
+                headers: rapina::extract::Headers,
+            ) -> String {
+                "ok".to_string()
+            }
+        };
+
+        let output = route_macro_core("GET", path, input);
+        let output_str = output.to_string();
+
+        // First two args use FromRequestParts
+        assert!(output_str.contains("FromRequestParts"));
+        // Last arg uses FromRequest (via blanket impl at runtime)
+        assert!(output_str.contains("FromRequest"));
+        // Request is destructured for multi-arg case
+        assert!(output_str.contains("into_parts"));
+        // Request is reassembled for last arg
+        assert!(output_str.contains("from_parts"));
+    }
+
+    #[test]
+    #[should_panic(expected = "expected function")]
+    fn test_invalid_input_panics() {
+        let path = quote!("/");
+        let invalid_input = quote! { not_a_function };
+
+        route_macro_core("GET", path, invalid_input);
+    }
+
+    #[test]
+    fn test_json_return_type_generates_response_schema() {
+        let path = quote!("/users");
+        let input = quote! {
+            async fn get_user() -> Json<UserResponse> {
+                Json(UserResponse { id: 1 })
+            }
+        };
+
+        let output = route_macro_core("GET", path, input);
+        let output_str = output.to_string();
+
+        // Check response_schema method is generated with openapi_schema_for
+        assert!(output_str.contains("fn response_schema"));
+        assert!(output_str.contains("rapina :: openapi_schema_for"));
+        assert!(output_str.contains("UserResponse"));
+    }
+
+    #[test]
+    fn test_result_json_return_type_generates_response_schema() {
+        let path = quote!("/users");
+        let input = quote! {
+            async fn get_user() -> Result<Json<UserResponse>> {
+                Ok(Json(UserResponse { id: 1 }))
+            }
+        };
+
+        let output = route_macro_core("GET", path, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("fn response_schema"));
+        assert!(output_str.contains("rapina :: openapi_schema_for"));
+        assert!(output_str.contains("UserResponse"));
+    }
+
+    #[test]
+    fn test_errors_attr_generates_error_responses() {
+        let path = quote!("/users");
+        let input = quote! {
+            #[errors(UserError)]
+            async fn get_user() -> Result<Json<UserResponse>> {
+                Ok(Json(UserResponse { id: 1 }))
+            }
+        };
+
+        let output = route_macro_core("GET", path, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("fn error_responses"));
+        assert!(output_str.contains("DocumentedError"));
+        assert!(output_str.contains("UserError"));
+    }
+
+    #[test]
+    fn test_json_body_generates_request_schema_and_content_type() {
+        let path = quote!("/users");
+        let input = quote! {
+            async fn create_user(body: Json<CreateUserRequest>) -> Json<UserResponse> {
+                Json(UserResponse { id: 1 })
+            }
+        };
+
+        let output = route_macro_core("POST", path, input);
+        let output_str = output.to_string();
+
+        // Check request_schema method is generated
+        assert!(output_str.contains("fn request_schema"));
+        assert!(output_str.contains("CreateUserRequest"));
+        // Check request_content_type method is generated with JSON content type
+        assert!(output_str.contains("fn request_content_type"));
+        assert!(output_str.contains("application/json"));
+    }
+
+    #[test]
+    fn test_form_body_generates_request_schema_and_content_type() {
+        let path = quote!("/users");
+        let input = quote! {
+            async fn create_user(body: Form<CreateUserForm>) -> Json<UserResponse> {
+                Json(UserResponse { id: 1 })
+            }
+        };
+
+        let output = route_macro_core("POST", path, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("fn request_schema"));
+        assert!(output_str.contains("CreateUserForm"));
+        // Check request_content_type method is generated with form content type
+        assert!(output_str.contains("fn request_content_type"));
+        assert!(output_str.contains("application/x-www-form-urlencoded"));
+    }
+
+    #[test]
+    fn test_validated_json_generates_request_schema_and_content_type() {
+        let path = quote!("/users");
+        let input = quote! {
+            async fn create_user(body: Validated<Json<CreateUserRequest>>) -> Json<UserResponse> {
+                Json(UserResponse { id: 1 })
+            }
+        };
+
+        let output = route_macro_core("POST", path, input);
+        let output_str = output.to_string();
+
+        // Should extract CreateUserRequest from Validated<Json<CreateUserRequest>>
+        assert!(output_str.contains("fn request_schema"));
+        assert!(output_str.contains("CreateUserRequest"));
+        // Should inherit JSON content type from inner Json extractor
+        assert!(output_str.contains("fn request_content_type"));
+        assert!(output_str.contains("application/json"));
+    }
+
+    #[test]
+    fn test_validated_form_generates_request_schema_and_content_type() {
+        let path = quote!("/login");
+        let input = quote! {
+            async fn login(body: Validated<Form<LoginForm>>) -> Json<TokenResponse> {
+                Json(TokenResponse { token: "abc".into() })
+            }
+        };
+
+        let output = route_macro_core("POST", path, input);
+        let output_str = output.to_string();
+
+        // Should extract LoginForm from Validated<Form<LoginForm>>
+        assert!(output_str.contains("fn request_schema"));
+        assert!(output_str.contains("LoginForm"));
+        // Should inherit form content type from inner Form extractor
+        assert!(output_str.contains("fn request_content_type"));
+        assert!(output_str.contains("application/x-www-form-urlencoded"));
+    }
+
+    #[test]
+    fn test_option_json_generates_optional_request_body() {
+        let path = quote!("/users");
+        let input = quote! {
+            async fn update_user(body: Option<Json<UpdateUserRequest>>) -> Json<UserResponse> {
+                Json(UserResponse { id: 1 })
+            }
+        };
+
+        let output = route_macro_core("PATCH", path, input);
+        let output_str = output.to_string();
+
+        // Should extract UpdateUserRequest from Option<Json<UpdateUserRequest>>
+        assert!(output_str.contains("fn request_schema"));
+        assert!(output_str.contains("UpdateUserRequest"));
+        // Should have JSON content type
+        assert!(output_str.contains("fn request_content_type"));
+        assert!(output_str.contains("application/json"));
+        // Should have request_body_required returning false
+        assert!(output_str.contains("fn request_body_required"));
+        assert!(output_str.contains("Some (false)"));
+    }
+
+    #[test]
+    fn test_option_form_generates_optional_request_body() {
+        let path = quote!("/login");
+        let input = quote! {
+            async fn login(body: Option<Form<LoginForm>>) -> Json<TokenResponse> {
+                Json(TokenResponse { token: "abc".into() })
+            }
+        };
+
+        let output = route_macro_core("POST", path, input);
+        let output_str = output.to_string();
+
+        // Should extract LoginForm from Option<Form<LoginForm>>
+        assert!(output_str.contains("fn request_schema"));
+        assert!(output_str.contains("LoginForm"));
+        // Should have form content type
+        assert!(output_str.contains("fn request_content_type"));
+        assert!(output_str.contains("application/x-www-form-urlencoded"));
+        // Should have request_body_required returning false
+        assert!(output_str.contains("fn request_body_required"));
+        assert!(output_str.contains("Some (false)"));
+    }
+
+    #[test]
+    fn test_get_with_json_body_no_request_schema() {
+        // GET handlers should not generate requestBody even if they have Json<T> parameter
+        let path = quote!("/users");
+        let input = quote! {
+            async fn list_users(body: Json<FilterRequest>) -> Json<Vec<UserResponse>> {
+                Json(vec![])
+            }
+        };
+
+        let output = route_macro_core("GET", path, input);
+        let output_str = output.to_string();
+
+        // Should NOT generate request_schema for GET method
+        assert!(!output_str.contains("fn request_schema"));
+        assert!(!output_str.contains("fn request_content_type"));
+        assert!(!output_str.contains("fn request_body_required"));
+    }
+
+    #[test]
+    fn test_delete_with_json_body_no_request_schema() {
+        // DELETE handlers should not generate requestBody even if they have Json<T> parameter
+        let path = quote!("/users/:id");
+        let input = quote! {
+            async fn delete_user(body: Json<DeleteRequest>) -> StatusCode {
+                StatusCode::NO_CONTENT
+            }
+        };
+
+        let output = route_macro_core("DELETE", path, input);
+        let output_str = output.to_string();
+
+        // Should NOT generate request_schema for DELETE method
+        assert!(!output_str.contains("fn request_schema"));
+        assert!(!output_str.contains("fn request_content_type"));
+        assert!(!output_str.contains("fn request_body_required"));
+    }
+
+    #[test]
+    fn test_no_body_no_request_schema_or_content_type() {
+        let path = quote!("/users");
+        let input = quote! {
+            async fn list_users() -> Json<Vec<UserResponse>> {
+                Json(vec![])
+            }
+        };
+
+        let output = route_macro_core("GET", path, input);
+        let output_str = output.to_string();
+
+        // Should NOT generate request_schema or request_content_type for handlers without body
+        assert!(!output_str.contains("fn request_schema"));
+        assert!(!output_str.contains("fn request_content_type"));
+    }
+
+    #[test]
+    fn test_non_json_return_type_no_response_schema() {
+        let path = quote!("/health");
+        let input = quote! {
+            async fn health() -> &'static str {
+                "ok"
+            }
+        };
+
+        let output = route_macro_core("GET", path, input);
+        let output_str = output.to_string();
+
+        // Check response_schema method is NOT generated for non-Json types
+        assert!(!output_str.contains("fn response_schema"));
+        assert!(!output_str.contains("openapi_schema_for"));
+    }
+
+    #[test]
+    fn test_user_state_variable_not_shadowed() {
+        // Regression test for issue #134 - user naming their extractor 'state'
+        // should not conflict with internal macro variables
+        let path = quote!("/users");
+        let input = quote! {
+            async fn list_users(state: rapina::extract::State<MyState>) -> String {
+                "ok".to_string()
+            }
+        };
+
+        let output = route_macro_core("GET", path, input);
+        let output_str = output.to_string();
+
+        // Internal variables should use __rapina_ prefix
+        assert!(output_str.contains("__rapina_state"));
+        assert!(output_str.contains("__rapina_params"));
+        // User's variable 'state' should still be extracted
+        assert!(output_str.contains("let state ="));
+    }
+
+    #[test]
+    fn test_no_closure_wrapper_for_type_inference() {
+        // Regression test for issue #134 - Result type inference should work
+        let path = quote!("/users");
+        let input = quote! {
+            async fn get_user() -> Result<String, Error> {
+                Ok("user".to_string())
+            }
+        };
+
+        let output = route_macro_core("GET", path, input);
+        let output_str = output.to_string();
+
+        // Should NOT use closure wrapper (|| async ...)
+        assert!(!output_str.contains("|| async"));
+        // Should use typed result with async block (: ReturnType = (async ...).await)
+        assert!(output_str.contains("__rapina_result"));
+        assert!(output_str.contains("Result < String , Error >"));
+    }
+
+    #[test]
+    fn test_emits_route_descriptor() {
+        let path = quote!("/users");
+        let input = quote! {
+            async fn list_users() -> &'static str {
+                "users"
+            }
+        };
+
+        let output = route_macro_core("GET", path, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("inventory :: submit !"));
+        assert!(output_str.contains("RouteDescriptor"));
+        assert!(output_str.contains("method : \"GET\""));
+        assert!(output_str.contains("path : \"/users\""));
+        assert!(output_str.contains("handler_name : \"list_users\""));
+        assert!(output_str.contains("is_public : false"));
+        assert!(output_str.contains("__rapina_register_list_users"));
+    }
+
+    #[test]
+    fn test_emits_route_descriptor_with_method() {
+        let path = quote!("/users");
+        let input = quote! {
+            async fn create_user() -> &'static str {
+                "created"
+            }
+        };
+
+        let output = route_macro_core("POST", path, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("method : \"POST\""));
+        assert!(output_str.contains("__rapina_router . post"));
+    }
+
+    #[test]
+    fn test_public_attr_below_route_sets_is_public() {
+        let path = quote!("/health");
+        let input = quote! {
+            #[public]
+            async fn health() -> &'static str {
+                "ok"
+            }
+        };
+
+        let output = route_macro_core("GET", path, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("is_public : true"));
+    }
+
+    #[test]
+    fn test_cache_attr_injects_ttl_header() {
+        let path = quote!("/products");
+        let input = quote! {
+            #[cache(ttl = 60)]
+            async fn list_products() -> &'static str {
+                "products"
+            }
+        };
+
+        let output = route_macro_core("GET", path, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("x-rapina-cache-ttl"));
+        assert!(output_str.contains("60"));
+    }
+
+    #[test]
+    fn test_no_cache_attr_no_ttl_header() {
+        let path = quote!("/products");
+        let input = quote! {
+            async fn list_products() -> &'static str {
+                "products"
+            }
+        };
+
+        let output = route_macro_core("GET", path, input);
+        let output_str = output.to_string();
+
+        assert!(!output_str.contains("x-rapina-cache-ttl"));
+    }
+
+    #[test]
+    fn test_cache_attr_with_extractors() {
+        let path = quote!("/users/:id");
+        let input = quote! {
+            #[cache(ttl = 120)]
+            async fn get_user(id: rapina::extract::Path<u64>) -> String {
+                format!("{}", id.into_inner())
+            }
+        };
+
+        let output = route_macro_core("GET", path, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("x-rapina-cache-ttl"));
+        assert!(output_str.contains("120"));
+        // Single arg uses FromRequest (positional convention)
+        assert!(output_str.contains("FromRequest"));
+    }
+
+    #[test]
+    fn test_group_param_joins_path() {
+        let attr = quote!("/users", group = "/api");
+        let input = quote! {
+            async fn list_users() -> &'static str {
+                "users"
+            }
+        };
+
+        let output = route_macro_core("GET", attr, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("path : \"/api/users\""));
+        assert!(output_str.contains("__rapina_router . get (\"/api/users\""));
+    }
+
+    #[test]
+    fn test_group_param_with_nested_prefix() {
+        let attr = quote!("/items", group = "/api/v1");
+        let input = quote! {
+            async fn list_items() -> &'static str {
+                "items"
+            }
+        };
+
+        let output = route_macro_core("GET", attr, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("path : \"/api/v1/items\""));
+    }
+
+    #[test]
+    fn test_without_group_param_backward_compatible() {
+        let attr = quote!("/users");
+        let input = quote! {
+            async fn list_users() -> &'static str {
+                "users"
+            }
+        };
+
+        let output = route_macro_core("GET", attr, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("path : \"/users\""));
+        assert!(output_str.contains("__rapina_router . get (\"/users\""));
+    }
+
+    #[test]
+    #[should_panic(expected = "group prefix must start with `/`")]
+    fn test_group_prefix_must_start_with_slash() {
+        let attr = quote!("/users", group = "api");
+        let input = quote! {
+            async fn list_users() -> &'static str {
+                "users"
+            }
+        };
+
+        route_macro_core("GET", attr, input);
+    }
+
+    #[test]
+    fn test_group_with_trailing_slash_normalized() {
+        let attr = quote!("/users", group = "/api/");
+        let input = quote! {
+            async fn list_users() -> &'static str {
+                "users"
+            }
+        };
+
+        let output = route_macro_core("GET", attr, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("path : \"/api/users\""));
+    }
+
+    #[test]
+    fn test_group_with_public_attr() {
+        let attr = quote!("/health", group = "/api");
+        let input = quote! {
+            #[public]
+            async fn health() -> &'static str {
+                "ok"
+            }
+        };
+
+        let output = route_macro_core("GET", attr, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("path : \"/api/health\""));
+        assert!(output_str.contains("is_public : true"));
+    }
+
+    #[test]
+    fn test_group_with_cache_attr() {
+        let attr = quote!("/products", group = "/api");
+        let input = quote! {
+            #[cache(ttl = 60)]
+            async fn list_products() -> &'static str {
+                "products"
+            }
+        };
+
+        let output = route_macro_core("GET", attr, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("path : \"/api/products\""));
+        assert!(output_str.contains("x-rapina-cache-ttl"));
+        assert!(output_str.contains("60"));
+    }
+
+    #[test]
+    fn test_group_with_errors_attr() {
+        let attr = quote!("/users", group = "/api");
+        let input = quote! {
+            #[errors(UserError)]
+            async fn get_user() -> Result<Json<UserResponse>> {
+                Ok(Json(UserResponse { id: 1 }))
+            }
+        };
+
+        let output = route_macro_core("GET", attr, input);
+        let output_str = output.to_string();
+
+        assert!(output_str.contains("path : \"/api/users\""));
+        assert!(output_str.contains("fn error_responses"));
+        assert!(output_str.contains("UserError"));
+    }
+
+    #[test]
+    fn test_group_with_all_methods() {
+        for method in &["GET", "POST", "PUT", "DELETE"] {
+            let attr = quote!("/items", group = "/api");
+            let input = quote! {
+                async fn handler() -> &'static str {
+                    "ok"
+                }
+            };
+
+            let output = route_macro_core(method, attr, input);
+            let output_str = output.to_string();
+
+            assert!(
+                output_str.contains("path : \"/api/items\""),
+                "{method} should produce /api/items"
+            );
+            let method_lower = method.to_lowercase();
+            assert!(
+                output_str.contains(&format!("__rapina_router . {method_lower}")),
+                "{method} should use .{method_lower}() on router"
+            );
+        }
+    }
+
+    #[test]
+    fn test_join_paths_basic() {
+        assert_eq!(join_paths("/api", "/users"), "/api/users");
+        assert_eq!(join_paths("/api/v1", "/items"), "/api/v1/items");
+    }
+
+    #[test]
+    fn test_join_paths_trailing_slash() {
+        assert_eq!(join_paths("/api/", "/users"), "/api/users");
+    }
+
+    #[test]
+    fn test_join_paths_empty_path() {
+        assert_eq!(join_paths("/api", ""), "/api");
+        assert_eq!(join_paths("/api", "/"), "/api");
+    }
+
+    #[test]
+    fn test_join_paths_empty_prefix() {
+        assert_eq!(join_paths("", "/users"), "/users");
+        assert_eq!(join_paths("", ""), "/");
     }
 }
